Refactor imports for @hedvig-ui/Checkbox
